requirements.txt
Flask==1.0.3
Flask-MySQL==1.5.2 
PyMySQL==0.9.3
uWSGI==2.0.17.1
mysql-connector-python
cryptography